Add first draft of QueryCacheEngineBridge.

+++ src/main/java/org/joshsim/engine/simulation/Replicate.java
@@ -50,5 +50,5 @@ public interface Replicate {
    * @param key of the Patch to lookup.
    * @param stepNumber of the timestep at which to return the patch.
    */
-  Patch getPatchByKey(PatchKey key, int stepNumber);
+  Patch getPatchByKey(PatchKey key, long stepNumber);
 }
